+++ R/Lib_ImageProcess.R
@@ -4,6 +4,7 @@
 # ==============================================================================
 # PROGRAMMERS:
 # Jean-Baptiste FERET <jb.feret@irstea.fr>
+# Florian de Boissieu <fdeboiss@gmail.com>
 # Copyright 2018/07 Jean-Baptiste FERET
 # ==============================================================================
 # This Library contains functions to manipulate & process raster images
@@ -60,9 +61,7 @@ change_resolution_HDR <- function(HDR, window_size) {
   return(HDR)
 }
 
-# clean data matrix from inf values and identifies if some values
-# (columns) are constant. variables showing constant value
-# need to be eliminated before PCA
+# remove constant bands
 #
 # @param DataMatrix each variable is a column
 # @param Spectral summary of spectral information: which spectral bands selected from initial data
@@ -70,7 +69,7 @@ change_resolution_HDR <- function(HDR, window_size) {
 # @return updated DataMatrix and Spectral
 # ' @importFrom stats sd
 
-check_invariant_bands <- function(DataMatrix, Spectral) {
+rm_invariant_bands <- function(DataMatrix, Spectral) {
   # samples with inf value are eliminated
   for (i in 1:ncol(DataMatrix)) {
     elim <- which(DataMatrix[, i] == Inf)
@@ -78,7 +77,7 @@ check_invariant_bands <- function(DataMatrix, Spectral) {
       DataMatrix <- DataMatrix[-elim, ]
     }
   }
-  # bands showing null std are eliminated
+  # bands showing null std are removed
   stdsub <- apply(DataMatrix, 2, sd)
   BandsNoVar <- which(stdsub == 0)
   # BandsNoVar  = which(stdsub<=0.002)
@@ -203,6 +202,12 @@ exclude_spectral_domains <- function(ImPath, Excluded_WL = FALSE) {
       Excluded_WL <- rbind(Excluded_WL, c(1780, 2040))
     }
   }
+  message("Exclude spectral domains corresponding to water vapor absorption")
+  message("The following spectral domains (min and max spectral band in nanometers) will be discarded")
+  print(Excluded_WL)
+  message("Please define the input variable Excluded_WL when calling perform_PCA")
+  message("if you want to modify these spectral domains")
+
   # in case a unique specrtal domain is provided as excluded domain
   if (length(Excluded_WL)==2){
     Excluded_WL = matrix(Excluded_WL,ncol = 2)
@@ -210,6 +215,9 @@ exclude_spectral_domains <- function(ImPath, Excluded_WL = FALSE) {
   # get image header data
   ImPathHDR <- get_HDR_name(ImPath)
   HDR <- read_ENVI_header(ImPathHDR)
+  if ((HDR$`wavelength units`=='Micrometers') | (HDR$`wavelength units`=='micrometers')){
+    Excluded_WL <- 0.001*Excluded_WL
+  }
   nbchannels0 <- HDR$bands
   idOkBand <- seq(1, nbchannels0)
   if ("wavelength" %in% names(HDR)) {
@@ -301,15 +309,15 @@ extract_samples_from_image <- function(ImPath, coordPix, MaxRAM = FALSE, Already
     if (dim(coordPix)[2] == 2) {
       if (dim(coordPix)[1] >= 2) {
         coordPix_tmp <- list()
-        coordPix_tmp$Row <- coordPix[, 1]
-        coordPix_tmp$Column <- coordPix[, 2]
+        coordPix_tmp$row <- coordPix[, 1]
+        coordPix_tmp$col <- coordPix[, 2]
       } else if (dim(coordPix)[1] == 1) {
         coordPix_tmp <- list()
-        coordPix_tmp$Row <- coordPix[1]
-        coordPix_tmp$Column <- coordPix[2]
+        coordPix_tmp$row <- coordPix[1]
+        coordPix_tmp$col <- coordPix[2]
       }
     }
-  } else if (typeof(coordPix) == "list" & length(grep("Row", names(coordPix))) > 0 & length(grep("Column", names(coordPix))) > 0) {
+  } else if (typeof(coordPix) == "list" & length(grep("row", names(coordPix))) > 0 & length(grep("col", names(coordPix))) > 0) {
     coordPix_tmp <- coordPix
   }
   # initial index value of the pixels requested in the image, following original ranking
@@ -317,8 +325,8 @@ extract_samples_from_image <- function(ImPath, coordPix, MaxRAM = FALSE, Already
   # rank of the initial list of pixels
   Rank_Index_Init <- sort(Index_Init, index = TRUE)
   # convert
-  if (typeof(coordPix) == "list" & length(grep("Row", names(coordPix))) > 0 & length(grep("Column", names(coordPix))) > 0) {
-    coordPix <- cbind(coordPix$Row, coordPix$Column)
+  if (typeof(coordPix) == "list" & length(grep("row", names(coordPix))) > 0 & length(grep("col", names(coordPix))) > 0) {
+    coordPix <- cbind(coordPix$row, coordPix$col)
   }
   # divide data access based on the size of the image (to avoid reading 30 Gb file at once...)
   Image_Format <- ENVI_type2bytes(HDR)
@@ -354,16 +362,16 @@ extract_samples_from_image <- function(ImPath, coordPix, MaxRAM = FALSE, Already
 
   # re-sort samples
   Sample_Sort <- list()
-  Sample_Sort$Row <- coordPix_List[, 1]
-  Sample_Sort$Column <- coordPix_List[, 2]
+  Sample_Sort$row <- coordPix_List[, 1]
+  Sample_Sort$col <- coordPix_List[, 2]
   Coord_Pixels <- sub2ind(HDR, Sample_Sort)
   # rank of the pixels extracted from the image
   Rank_Pixels <- sort(Coord_Pixels, index = TRUE)
   # pixel re-ordering needs to be performed in order to get back to Rank_Index_Init
   # first re-order pixels in order to follow increasing index value
   # Sample_Sort$index   = Coord_Pixels[Rank_Pixels$ix]
-  # Sample_Sort$Row     = Sample_Sort$Row[Rank_Pixels$ix]
-  # Sample_Sort$Column  = Sample_Sort$Column[Rank_Pixels$ix]
+  # Sample_Sort$row     = Sample_Sort$row[Rank_Pixels$ix]
+  # Sample_Sort$col  = Sample_Sort$col[Rank_Pixels$ix]
   # if bug, check this line
   # Sample_Sel          = Sample_Sel[Rank_Pixels$ix]
   Sample_Sel <- Sample_Sel[Rank_Pixels$ix, ]
@@ -377,36 +385,106 @@ extract_samples_from_image <- function(ImPath, coordPix, MaxRAM = FALSE, Already
   return(Sample_Sel)
 }
 
-# Perform random sampling on an image including a mask
+# Extract bands of sparse pixels in big image data cube.
 #
-# @param ImPath path for image
-# @param HDR path for hdr file
-# @param MaskPath path for mask
-# @param nb_partitions number of k-means then averaged
-# @param Pix_Per_Partition number of pixels per iteration
+# Extract bands of sparse pixels in big image data cube, typically hyperspectral data cube.
+# @param ImPath character. Path to the image cube
+# @param rowcol matrix or data.frame with two columns: row, col.
+# If columns are not named, 1st=row, 2nd=col.
+# @param MaxRAM numeric. Maximum memory use at block reading.
+# It constrains the maximum number rows of a block
 #
-# @return samples from image, coordinates of these samples and updated number of pixels to sample
+# @return matrix. Rows are corresponding to the samples, columns are the bands.
+#' @import stars
+extract.big_raster <- function(ImPath, rowcol, MaxRAM=.25){
+  # ImPath = hs_file
+  # rowcol = arcd
+  # rowcol = as.data.table(rowcol)
+  # MaxRAM = .25
+
+  if(!is.data.frame(rowcol)){
+    rowcol <- as.data.frame(rowcol)
+  }
+
+  if(!all(c('row', 'col') %in% colnames(rowcol))){
+    warning('Columns row,col not found in rowcol argument. The two first columns are considered as row, col respectively.')
+    colnames(rowcol)[1:2]= c('row', 'col')
+  }
+
+
+  r = brick(ImPath)
+  # nbytes = as.numeric(substring(dataType(r), 4, 4))
+  # stars converts automatically values to numeric
+  nbytes = 8
+  ImgSizeGb = prod(dim(r))*nbytes/2^30
+  LineSizeGb = prod(dim(r)[2:3])*nbytes/2^30
+  LinesBlock = floor(MaxRAM/LineSizeGb)
+  rowcol$rowInBlock = ((rowcol$row-1) %% LinesBlock)+1  # row in block number
+  rowcol$block=floor((rowcol$row-1)/LinesBlock)+1  # block number
+  rowcol$sampleIndex = 1:nrow(rowcol)  # sample index to reorder result
+
+  sampleList = lapply(unique(rowcol$block), function(iblock){
+    rc = rowcol[rowcol$block==iblock,]
+    rr = range(rc$row)
+    nYSize = diff(rr)+1
+    nXSize = max(rc$col)
+    # stars data cube dimension order is x*y*band
+    ipix_stars = (rc$rowInBlock-min(rc$rowInBlock))*nXSize+rc$col
+    values = read_stars(ImPath, RasterIO =list(nXSize=nXSize, nYOff=rr[1], nYSize=nYSize))[[1]]
+    values = matrix(values, nrow=nYSize*nXSize)
+    res = cbind(rc$sampleIndex, values[ipix_stars, ])
+    rm('values')
+    gc()
+    return(res)
+  })
+
+  samples = do.call(rbind, sampleList)
+  samples = samples[order(samples[,1]),2:ncol(samples)]
+
+  return(samples)
+}
+
+# @return list, see Details
+# @details
+# The returned list contains:
+# - DataSubset: matrix of NxP of N samples and P bands
+# - nbPix2Sample: integer giving the number of pixels sampled (only central pixel of kernel)
+# - coordPix: a data.table with columns 'row', 'col' of pixel in the image corresponding to each row of DataSubset, and if kernel is not NULL
+# Kind (Kernel index) and 'id' the sample ID to be used with the kernel
 #' @importFrom matlab ones
-get_random_subset_from_image <- function(ImPath, HDR, MaskPath, nb_partitions, Pix_Per_Partition) {
+#' @import raster
+#' @importFrom mmand erode
+#' @importFrom data.table data.table rbindlist setorder
+#' @importFrom matrixStats rowAnys
+get_random_subset_from_image <- function(ImPath, HDR, MaskPath, nb_partitions, Pix_Per_Partition, kernel=NULL) {
+  r <- brick(ImPath)
   nbPix2Sample <- nb_partitions * Pix_Per_Partition
   # get total number of pixels
-  nbpix <- as.double(HDR$lines) * as.double(HDR$samples)
+  rdim <- dim(r)
+  nlines <- rdim[1]
+  nsamples <- rdim[2]
+  nbpix <- ncell(r)
   # 1- Exclude masked pixels from random subset
   # Read Mask
   if ((!MaskPath == "") & (!MaskPath == FALSE)) {
-    fid <- file(
-      description = MaskPath, open = "rb", blocking = TRUE,
-      encoding = getOption("encoding"), raw = FALSE
-    )
-    ShadeMask <- readBin(fid, integer(), n = nbpix, size = 1)
-    close(fid)
-    ShadeMask <- aperm(array(ShadeMask, dim = c(HDR$samples, HDR$lines)))
+    mask <- matrix(raster(MaskPath),ncol= HDR$samples,nrow = HDR$lines)
+    if(any(dim(mask)[1:2] != dim(r)[1:2])){
+      stop('Mask and Image rasters do not have the same XY diemnsions.')
+    }
   } else {
-    ShadeMask <- array(ones(HDR$lines * HDR$samples, 1), dim = c(HDR$lines, HDR$samples))
+    mask <- array(1, dim = c(nlines, nsamples))
   }
+
+  if(is.matrix(kernel)){
+    # erode mask with kernel, to keep valid central pixels and neighbours
+    mask = matlab::padarray(mask, c(1,1), padval=0, direction='both')
+    mask = erode(mask, (kernel!=0)*1)
+    mask = mask[2:(nrow(mask)-1), 2:(ncol(mask)-1)]
+  }
+
   # get a list of samples among unmasked pixels
-  IndexInit <- (matrix(1:nbpix, ncol = HDR$samples))
-  ValidPixels <- sample(which(ShadeMask == 1 | ShadeMask == 255))
+
+  ValidPixels <- which(mask > 0)
   NbValidPixels <- length(ValidPixels)
   # Check if number of valid pixels is compatible with number of pixels to be extracted
   # if number of pixels to sample superior to number of valid pixels, then adjust iterations
@@ -416,24 +494,57 @@ get_random_subset_from_image <- function(ImPath, HDR, MaskPath, nb_partitions, P
     Pix_Per_Partition <- floor(NbValidPixels / nb_partitions)
     nbPix2Sample <- nb_partitions * Pix_Per_Partition
   }
-  # Select a subset of nbPix2Sample among pixselected
-  pixselected <- ValidPixels[1:nbPix2Sample]
 
-  # define location of pixselected in binary file (avoid multiple reads) and optimize access to disk
-  # the file is a BIL file, which means that for each line in the image,
-  # we first need to define if datapoints are on the line, then read one point after the other
-  coordi <- ((pixselected - 1) %% HDR$lines) + 1
-  coordj <- floor((pixselected - 1) / HDR$lines) + 1
-  # sort based on line and col
-  indxPix <- order(coordi, coordj, na.last = FALSE)
-  coordPix <- cbind(coordi[indxPix], coordj[indxPix])
+  # Select a random subset of nbPix2Sample
+  seed <- sample(10000)[1]
+  set.seed(seed)
+  pixselected <- sample(ValidPixels, nbPix2Sample)
+  Row <- ((pixselected - 1) %% nlines) + 1 # row
+  Column <- floor((pixselected - 1) / nlines) + 1 # column
+  if(is.matrix(kernel)){
+    coordPixK = list()
+    mesh=matlab::meshgrid(-(ncol(kernel)%/%2):(ncol(kernel)%/%2), -(nrow(kernel)%/%2):(nrow(kernel)%/%2))
+    for(p in which(kernel!=0)){
+      coordPixK[[p]] = data.table(row = Row+mesh$y[p], col = Column+mesh$x[p], id=1:length(Row))
+    }
+    coordPix = rbindlist(coordPixK, idcol='Kind')
+    # Order along coordPix$id for further use in noise, mnf
+    setorder(coordPix, 'id')
+  }else{
+    coordPix = data.table(row = Row, col = Column, id = 1:length(Row))
+  }
+  # sort based on .bil dim order, i.e. band.x.y or band.col.row
+  # TODO: sorting may not be necessary anymore, neither unique coordinates
+  # setorder(coordPix, col, row)
+
+  # make unique
+  ucoordPix <- unique(coordPix[,c('row','col')])
+  ucoordPix[['sampleIndex']] = 1:nrow(ucoordPix)
 
   # 2- Extract samples from image
-  Sample_Sel <- extract_samples_from_image(ImPath, coordPix)
-  # randomize!
-  Sample_Sel <- Sample_Sel[sample(dim(Sample_Sel)[1]), ]
-  coordPix <- coordPix[sample(dim(Sample_Sel)[1]), ]
-  my_list <- list("DataSubset" = Sample_Sel, "nbPix2Sample" = nbPix2Sample,"coordPix"=coordPix)
+  # TODO: if coordPix can be a data.frame it would be easier
+  # Sample_Sel <- biodivMapR:::extract_samples_from_image(ImPath, ucoordPix)
+  Sample_Sel <- extract.big_raster(ImPath, ucoordPix[,1:2])
+  samplePixIndex = merge(coordPix, ucoordPix, by=c('row', 'col'), sort = FALSE)
+  # randomize! It should already be random from the sample operation on mask valid pixels
+  Sample_Sel <- Sample_Sel[samplePixIndex$sampleIndex, ]
+  samplePixIndex[['sampleIndex']]=NULL
+
+  # remove NA and Inf pixels
+  if(any(is.na(Sample_Sel) | is.infinite(Sample_Sel))){
+    print('Removing pixels with NA values.')
+    rmrows <- !rowAnys(is.na(Sample_Sel) | is.infinite(Sample_Sel))
+    rmpix <- unique(samplePixIndex$id[rmrows])
+    Sample_Sel = Sample_Sel[!(samplePixIndex$id %in% rmpix),]
+    samplePixIndex = samplePixIndex[!(samplePixIndex$id %in% rmpix)]
+    nbPix2Sample = length(unique(samplePixIndex$id))
+  }
+
+  ### FLORIAN: REMOVED RANDOMIZATION AS IT SEEMED USELESS
+  # Sample_Sel <- Sample_Sel[sample(dim(Sample_Sel)[1]), ]
+  # coordPix <- coordPix[sample(dim(Sample_Sel)[1]), ]
+  # TODO: check how returned coordPix is used, as it is now a data.frame
+  my_list <- list("DataSubset" = Sample_Sel, "nbPix2Sample" = nbPix2Sample,"coordPix"=samplePixIndex)
   return(my_list)
 }
 
@@ -525,18 +636,31 @@ ind2sub2 <- function(Raster, Image_Index) {
 #
 # @return rank of all spectral bands of interest in the image and corresponding wavelength
 #' @importFrom matlab padarray
-mean_filter <- function(ImageInit, nbi, nbj, SizeFilt) {
-  E <- padarray(ImageInit, c(SizeFilt, SizeFilt), "symmetric", "both")
-  ImageSmooth <- matrix(0, nrow = nbi, ncol = nbj)
-  Mat2D <- MatSun <- matrix(0, nrow = ((2 * SizeFilt) + 1)^2, ncol = nbj)
-  spl <- split(1:nbj, 1:((2 * SizeFilt) + 1))
-  mid <- ceiling((((2 * SizeFilt) + 1)^2) / 2)
-  for (i in (SizeFilt + 1):(nbi + SizeFilt)) {
-    for (j in 1:((2 * SizeFilt) + 1)) {
-      # create a 2D matrix
-      Mat2D[, spl[[j]]] <- matrix(E[(i - SizeFilt):(i + SizeFilt), (spl[[j]][1]):(tail(spl[[j]], n = 1) + 2 * SizeFilt)], nrow = ((2 * SizeFilt) + 1)^2)
+mean_filter <- function(Image, SizeFilt,NA_remove = FALSE) {
+  nbi <- dim(Image)[1]
+  nbj <- dim(Image)[2]
+  # if matrix: convert into array
+  if (is.na(dim(Image)[3])){
+    Image <- array(Image,dim = c(nbi,nbj,1))
+  }
+  ImageSmooth <- array(0,c(nbi,nbj,dim(Image)[3]))
+  for (band in 1: dim(Image)[3]){
+    E <- padarray(Image[,,band], c(SizeFilt, SizeFilt), "symmetric", "both")
+    ImageSmooth_tmp <- matrix(0, nrow = nbi, ncol = nbj)
+    Mat2D <- MatSun <- matrix(0, nrow = ((2 * SizeFilt) + 1)^2, ncol = nbj)
+    spl <- split(1:nbj, 1:((2 * SizeFilt) + 1))
+    mid <- ceiling((((2 * SizeFilt) + 1)^2) / 2)
+    for (i in (SizeFilt + 1):(nbi + SizeFilt)) {
+      for (j in 1:((2 * SizeFilt) + 1)) {
+        # create a 2D matrix
+        Mat2D[, spl[[j]]] <- matrix(E[(i - SizeFilt):(i + SizeFilt), (spl[[j]][1]):(tail(spl[[j]], n = 1) + 2 * SizeFilt)], nrow = ((2 * SizeFilt) + 1)^2)
+      }
+      ImageSmooth_tmp[(i - SizeFilt), ] <- colMeans(Mat2D, na.rm = TRUE)
     }
-    ImageSmooth[(i - SizeFilt), ] <- colMeans(Mat2D, na.rm = TRUE)
+    if (NA_remove == TRUE){
+      ImageSmooth_tmp[which(is.na(Image[,,band]))] <- NA
+    }
+    ImageSmooth[,,band] <- ImageSmooth_tmp
   }
   return(ImageSmooth)
 }
@@ -645,6 +769,8 @@ read_ENVI_header <- function(HDRpath) {
 # @param ImBand Bands to be read
 #
 # @return Image_Subset information corresponding to ImBand
+#' @import stars
+
 read_image_bands <- function(ImPath, HDR, ImBand) {
   # first get image format
   Image_Format <- ENVI_type2bytes(HDR)
@@ -652,42 +778,45 @@ read_image_bands <- function(ImPath, HDR, ImBand) {
   jpix <- HDR$samples
   nbChannels <- HDR$bands
   nbSubset <- length(ImBand)
-  # then open and read image
-  # depending on image size, need to read in one or multiple times
-  lenTot <- as.double(ipix) * as.double(jpix) * as.double(nbChannels)
-  lenSubset <- as.double(ipix) * as.double(jpix) * as.double(nbSubset)
-  ImSizeGb <- (lenTot * Image_Format$Bytes) / (1024^3)
-  # maximum image size read at once. If image larger, then reads in multiple pieces
-  LimitSizeGb <- 0.25
-  if (ImSizeGb < LimitSizeGb) {
-    nbLinesPerCPU <- ceiling(ipix)
-    nbPieces <- 1
-  } else {
-    # nb of lines corresponding to LimitSizeGb
-    OneLine <- as.double(jpix) * as.double(nbChannels) * Image_Format$Bytes
-    nbLinesPerCPU <- floor(LimitSizeGb * (1024^3) / OneLine)
-    # number of pieces to split the image into
-    nbPieces <- ceiling(ipix / nbLinesPerCPU)
+  Image_Subset <- array(NA,c(ipix,jpix,length(ImBand)))
+  i <- 0
+  for (band in ImBand){
+    i <- i+1
+    bndtmp = t(matrix(raster(ImPath, band = ImBand[i]),nrow = HDR$samples,ncol = HDR$lines))
+    Image_Subset[,,i] <- array(bndtmp,c(ipix,jpix,1))
   }
-  # Define segments of image to be read
-  SeqRead.Image <- where_to_read(HDR, nbPieces)
-  # Read segments (subsets) of image
-  Image_Subsets <- list()
-  for (i in 1:nbPieces) {
-    # number of elements to be read
-    Byte_Start <- SeqRead.Image$ReadByte_Start[i]
-    nbLines <- SeqRead.Image$Lines_Per_Chunk[i]
-    lenBin <- SeqRead.Image$ReadByte_End[i] - SeqRead.Image$ReadByte_Start[i] + 1
-    Image_Subsets[[i]] <- read_bin_subset(Byte_Start, nbLines, lenBin, ImPath, ImBand, jpix, nbChannels, Image_Format)
-  }
-  # reshape image with original size and selected bands
-  Image_Subset <- build_image_from_list(Image_Subsets, ipix, jpix, nbSubset)
-  rm(Image_Subsets)
-  gc()
   return(Image_Subset)
 }
 
-# reads subset of an image
+# reads subset of lines from an image
+#
+# @param ImPath path for the image
+# @param HDR header information corresponding to the image
+# @param Line_Start which line to start reading
+# @param Lines_To_Read number of lines to read
+#
+# @return data corresponding to the subset in original 3D format
+read_image_subset <- function(ImPath, HDR, Line_Start,Lines_To_Read,ImgFormat='3D'){
+  # list of pixels to be extracted
+  ListRows <- seq(Line_Start,Line_Start+Lines_To_Read-1)
+  ListCols <- seq(1,HDR$samples)
+  # list pixels
+  coordPix <- expand.grid(ListRows,ListCols)
+  names(coordPix) <- c('row','col')
+  coordPix[['sampleIndex']] = 1:nrow(coordPix)
+  # 2- Extract samples from image
+  Sample_Sel <- extract.big_raster(ImPath, coordPix[,1:2])
+  Sample_Sel <- array(matrix(as.numeric(unlist(Sample_Sel)),ncol = HDR$bands),c(Lines_To_Read,HDR$samples,HDR$bands))
+  if (ImgFormat == "2D") {
+    Sample_Sel <- array(Sample_Sel, c(Lines_To_Read * HDR$samples, HDR$bands))
+  }
+  if (ImgFormat == "Shade") {
+    Sample_Sel <- matrix(Sample_Sel, Lines_To_Read, HDR$samples, byrow = F)
+  }
+  return(Sample_Sel)
+}
+
+# reads subset of an ENVI BIL image
 #
 # @param ImPath path for the image
 # @param HDR header information corresponding to the image
@@ -698,7 +827,7 @@ read_image_bands <- function(ImPath, HDR, ImBand) {
 # @param ImgFormat should output be 2D or 3D (original image format)?
 #
 # @return data corresponding to the subset in original 3D format
-read_image_subset <- function(ImPath, HDR, Byte_Start, lenBin, nbLines, Image_Format, ImgFormat) {
+read_BIL_image_subset <- function(ImPath, HDR, Byte_Start, lenBin, nbLines, Image_Format, ImgFormat) {
   fidIm <- file(
     description = ImPath, open = "rb", blocking = TRUE,
     encoding = getOption("encoding"), raw = FALSE
@@ -843,6 +972,7 @@ update_shademask <- function(MaskPath, HDR, Mask, MaskPath_Update) {
   HDR_Update$`band names` <- {
     "Mask"
   }
+  HDR_Update$`default stretch` <- '0 1 linear'
   HDR_Update$wavelength <- NULL
   HDR_Update$fwhm <- NULL
   HDR_Update$resolution <- NULL
@@ -984,7 +1114,11 @@ Write_Big_Image <- function(ImgWrite,ImagePath,HDR,Image_Format){
     }
     ImgChunk <- aperm(ImgChunk, c(2, 3, 1))
     # writeBin(as.numeric(c(PCA_Chunk)), fidOUT, size = PCA_Format$Bytes,endian = .Platform$endian)
-    writeBin(c(ImgChunk), fidOUT, size = Image_Format$Bytes, endian = .Platform$endian, useBytes = FALSE)
+    if (!Image_Format$Bytes==1){
+      writeBin(c(ImgChunk), fidOUT, size = Image_Format$Bytes, endian = .Platform$endian, useBytes = FALSE)
+    } else {
+      writeBin(c(as.integer(ImgChunk)), fidOUT, size = Image_Format$Bytes, endian = .Platform$endian, useBytes = FALSE)
+    }
     close(fidOUT)
   }
   rm(ImgWrite)
@@ -993,6 +1127,102 @@ Write_Big_Image <- function(ImgWrite,ImagePath,HDR,Image_Format){
   return("")
 }
 
+#' write an image resulting from "window processing" at native spatial resolution
+#' (assuming square windows & origin at top left corner)
+#'
+#' @param Image numeric. Image corresponding to a 2D matrix or 3D array
+#' @param ImagePath character. Path where the image should be written
+#' @param HDR list. Image header
+#' @param window_size numeric. window size used to generate Image
+#'
+#' @return None
+#' @export
+
+Write_Image_NativeRes <- function(Image,ImagePath,HDR,window_size){
+  # update HDR: dimensions & spatial resolution
+  HDR_Full <- HDR
+  HDR_Full$samples <- HDR$samples * window_size
+  HDR_Full$lines <- HDR$lines * window_size
+  HDR_Full <- revert_resolution_HDR(HDR_Full, window_size)
+  # define name for native resolution image and corresponding HDR
+  ImagePath_FullRes <- paste(ImagePath, "_Fullres", sep = "")
+  headerFpath <- paste(ImagePath_FullRes, ".hdr", sep = "")
+  # write header
+  write_ENVI_header(HDR_Full, headerFpath)
+  Image_Format <- ENVI_type2bytes(HDR_Full)
+  # create image the same size as native image: convert matrix into array
+  if (length(dim(Image))==2){
+    Image = array(Image,c(HDR$lines,HDR$samples,1))
+  }
+  Image_FullRes <- array(NA,c(HDR_Full$lines,HDR_Full$samples,HDR_Full$bands))
+  for (band in  1:HDR_Full$bands){
+    for (i in 1:HDR$lines) {
+      for (j in 1:HDR$samples) {
+        Image_FullRes[((i-1)*window_size+1):(i*window_size),((j-1)*window_size+1):(j*window_size),band] <- Image[i,j,band]
+      }
+    }
+  }
+  # write image and make sure size does not matter ...
+  Write_Big_Image(ImgWrite = Image_FullRes,ImagePath = ImagePath_FullRes,
+                  HDR = HDR_Full,Image_Format = Image_Format)
+  # zip resulting file
+  ZipFile(ImagePath_FullRes)
+  return("")
+}
+
+
+# Writes a matrix or an array into a ENVI BIL raster
+#
+#' @param Image numeric. matrix or array of image to be written
+#' @param HDR hdr template
+#' @param ImagePath path of image file to be written
+#' @param window_size spatial units use dto compute diversiy (in pixels)
+#' @param FullRes should full resolution image be written (original pixel size)
+#' @param LowRes should low resolution image be written (one value per spatial unit)
+#
+#' @return None
+
+write_raster <- function(Image, HDR, ImagePath, window_size, FullRes = TRUE, LowRes = FALSE,SmoothImage = FALSE) {
+
+  Image_Format <- ENVI_type2bytes(HDR)
+  # Write image with resolution corresponding to window_size
+  if (LowRes == TRUE) {
+    # write header
+    headerFpath <- paste(ImagePath, ".hdr", sep = "")
+    write_ENVI_header(HDR, headerFpath)
+    # write image
+    Write_Big_Image(Image,ImagePath,HDR,Image_Format)
+  }
+
+  # Write image with Full native resolution
+  if (FullRes == TRUE) {
+    Write_Image_NativeRes(Image,ImagePath,HDR,window_size)
+  }
+
+  # write smoothed map
+  if (SmoothImage == TRUE){
+    SizeFilt <- 1
+    if (min(c(dim(Image)[1], dim(Image)[2])) > (2 * SizeFilt + 1)) {
+      Image_Smooth <- mean_filter(Image, SizeFilt,NA_remove = TRUE)
+      ImagePath.Smooth <- paste(ImagePath, "_MeanFilter", sep = "")
+      if (LowRes == TRUE) {
+        # write header
+        headerFpath <- paste(ImagePath.Smooth, ".hdr", sep = "")
+        write_ENVI_header(HDR, headerFpath)
+        # write image
+        Write_Big_Image(Image_Smooth,ImagePath.Smooth,HDR,Image_Format)
+        # close(fidOUT)
+      }
+      if (FullRes == TRUE) {
+        # Write image with Full native resolution
+        Write_Image_NativeRes(Image_Smooth,ImagePath.Smooth,HDR,window_size)
+      }
+    }
+  }
+  return("")
+}
+
+
 # convert image coordinates from X-Y to index
 #
 # @param HDR_Raster
@@ -1000,7 +1230,7 @@ Write_Big_Image <- function(ImgWrite,ImagePath,HDR,Image_Format){
 #
 # @return Image_Index
 sub2ind <- function(HDR_Raster, Pixels) {
-  Image_Index <- (Pixels$Column - 1) * HDR_Raster$lines + Pixels$Row
+  Image_Index <- (Pixels$col - 1) * HDR_Raster$lines + Pixels$row
   return(Image_Index)
 }
 
@@ -1050,12 +1280,12 @@ revert_resolution_HDR <- function(HDR, window_size) {
 # @param ImagePath path for the image
 # @return None
 ZipFile <- function(ImagePath) {
-  PathRoot <- getwd()
-  ImageDir <- dirname(ImagePath)
-  ImageName <- basename(ImagePath)
-  setwd(ImageDir)
-  zip::zip(zipfile = paste0(ImageName, ".zip"), files = ImageName)
-  file.remove(ImageName)
-  setwd(PathRoot)
+
+
+  ImagePath <- normalizePath(ImagePath)
+
+  zip::zipr(zipfile = paste0(ImagePath, ".zip"), files = ImagePath)
+  file.remove(ImagePath)
+
   return(invisible())
 }